Align/refactor `retryOnEmpty` & `block-ref` tests
These tests share similar test helpers and are written in similar
styles, so this commit extracts and consolidates those helpers and
tweaks the tests so they look more similar.
